Disentangle TEdgeConfig from `tedge config` CLI
